requirements.txt
torch==2.0.0
torchaudio==2.0.1
torchvision==0.15.1
pytorch-lightning==2.0.2
torchmetrics==0.11.4
scikit-learn==1.2.2
pandas==2.0.1
pytorch_metric_learning==2.1.2
hydra-core==1.3.2